+++ pages/signin/index.tsx
@@ -6,7 +6,7 @@ import classNames from "classnames/bind";
 import { SubmitHandler, useForm } from "react-hook-form";
 import { zodResolver } from "@hookform/resolvers/zod";
 import { schema, FormFields } from "@/lib/schema";
-import { LoginLayout } from "@/components/layout/login-layout/LoginLayout";
+import { LoginLayout } from "@/components/layout/page-layout/LoginLayout/LoginLayout";
 import Eyeoff from "@/public/Eyeoff.svg";
 import Eyeon from "@/public/Eyeon.svg";
 import SocialBox from "@/components/login-social-box/SocialBox";
@@ -78,6 +78,7 @@ export default function Signin() {
           placeholder="비밀번호를 입력해 주세요."
         />
         <button
+          tabIndex={-1}
           className={cx("eye-icon")}
           onClick={handleToggleClick}
           type="button"
